refactor(tests): consolidate orphaned web tests, ambient types, and utility folders
