+++ apps/web/src/lib/browse/getCategoryIcon.ts
